bme280: code restructured for bme280
