+++ src/bme280/CMakeLists.txt
@@ -1,6 +1,6 @@
 include_directories(..)
 set (libname "bme280")
 set (libdescription "upm bme280 temperature humidity pressure sensor module")
-set (module_src ${libname}.cxx)
-set (module_h ${libname}.h)
+set (module_src ${libname}.cxx bme280driver.cxx)
+set (module_h ${libname}.h bme280driver.h)
 upm_module_init()
